[move] move interface directory to under src directory
