+++ src/interfaces/imageConverter.ts
