+++ src/interfaces/imageStorage.ts
