+++ src/s3ImageStorage.ts
@@ -1,4 +1,4 @@
-import { IImageStorage } from '../interfaces/imageStorage';
+import { IImageStorage } from './interfaces/imageStorage';
 import { S3Client, PutObjectCommand, GetObjectCommand } from '@aws-sdk/client-s3';
 import { getSignedUrl } from '@aws-sdk/s3-request-presigner';
